reorganized things to auto-forward script variables to the OpenCL code.  broke ADM-1D in the process.  learned that toNumericString, which depends on std::to_string, uses res/config.luaf by default.  Time to find a better string/number formatter.

+++ include/HydroGPU/toNumericString.h
@@ -6,7 +6,7 @@
 template<typename T> std::string toNumericString(T value);
 
 template<> inline std::string toNumericString<double>(double value) {
-	std::string s = std::to_string(value);
+	std::string s = std::to_string((long double)value);
 	if (s.find("e") == std::string::npos) {
 		if (s.find(".") == std::string::npos) {
 			s += ".";
